Extract Expression into its own file.
